resource: Refactor Scoring API code
Improve readability by refactoring the scoring api code
into multiple files.

Move edge comparator classes into fold.hpp

Move edge evaluator classes into edge_eval_api.hpp and .cpp.

Move the definition of score_api_t methods into its own
source file, scoring_api.cpp (except for templated methods).
